pyproject.toml
[tool.poetry]
name = "dj-vach"
version = "0.1.0"
description = ""
authors = ["Sergey <one6sergey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
ruff = "^0.6.4"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-ninja = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
